Split St7789 driver into base and DMA versions. Instantiate for SoC Bootloader. SoC RAM to 192kb.
