+++ SoC/Rtl/SoC.srcs/sources_1/new/GpioBlock.sv
@@ -50,7 +50,7 @@ module GpioBlock #(parameter ADDRESS)(
         begin
             integer i;
             for (i = 0; i < IO_WORDS; i = i + 1) begin
-                MEM[i] = 0;
+                MEM[i] <= 0;
             end
         end
         else if (enabled)
