+++ SoC/Rtl/SoC.srcs/sources_1/new/Main.sv
@@ -65,21 +65,6 @@ module Main(
     wire soc_clk = (sw[15] ? manual_clk : cpu_clk);
 
     //Reset
-    /*
-    reg reset;
-    reg running;
-    always_ff @(posedge cpu_clk) begin
-        if (!running)
-            if (!reset) begin
-                reset <= 1;
-                running <=0;
-            end else begin
-                reset <= 0;
-                running <= 1;
-            end
-    end
-    */
-
     wire reset = sw[14];
 
     //SoC
