+++ SoC/Rtl/SoC.srcs/sources_1/new/Memory.sv
@@ -21,7 +21,7 @@
 
 `include "DEFINES.vinc"
 
-module Memory #(parameter ADDRESS, parameter WORDS = 32)(
+module Memory #(parameter ADDRESS, parameter ADDR_WIDTH) (
     //Memory interface
     input cpu_clk,
     input reset,
@@ -31,34 +31,26 @@ module Memory #(parameter ADDRESS, parameter WORDS = 32)(
     input [3:0] write
 );
 
-    (* ram_style = "block" *)
-    (* ram_decomp = "power" *)
-    reg [31:0] MEM [0:(WORDS - 1)];
-    reg [31:0] memData;
-    wire enabled = ((address >= ADDRESS[31:2]) && (address < (ADDRESS[31:2] + WORDS)));
+    localparam MAX_ADDR = 2**ADDR_WIDTH + ADDRESS;
+    wire enabled = ((address >= ADDRESS[31:2]) && (address < MAX_ADDR));
 
-    initial begin
-        $display("Loading RAM.");
-        $readmemh("fpga.mem", MEM);
-    end
+    wire [31:0] bramDataOut;
+    bytewrite_tdp_ram_rf #(.ADDR_WIDTH(ADDR_WIDTH)) block_ram(
+        .clkA(cpu_clk),
+        .reset(reset),
+        .enaA(enabled),
+        .weA(write),
+        .addrA(address[ADDR_WIDTH+1:2]),
+        .dinA(dataIn),
+        .doutA(bramDataOut)
+        //.clkB(),
+        //.enaB(),
+        //.weB(),
+        //.addrB(),
+        //.dinB(),
+        //.doutB()
+    );
 
-    wire [29:0] wordAddress = address[31:2];
-
-    always_ff @(posedge cpu_clk) begin
-        if (reset)
-        begin
-            $readmemh("fpga.mem", MEM);
-        end
-        else if (enabled)
-        begin
-            if(write[0]) MEM[wordAddress][7:0] <= dataIn[7:0];
-            if(write[1]) MEM[wordAddress][15:8] <= dataIn[15:8];
-            if(write[2]) MEM[wordAddress][23:16] <= dataIn[23:16];
-            if(write[3]) MEM[wordAddress][31:24] <= dataIn[31:24];
-            memData <= MEM[wordAddress];
-        end
-    end
-
-    assign dataOut = enabled ? memData : 32'hzzzzzzzz;
+    assign dataOut = enabled ? bramDataOut : 32'hzzzzzzzz;
 
 endmodule
