+++ Src/SoC/Bootloader/linker.ld
@@ -2,7 +2,7 @@ ENTRY(_enter)
 
 MEMORY
 {
-   BRAM (RWX) : ORIGIN = 0x0000, LENGTH = 0x10000 /* 64k */
+   BRAM (RWX) : ORIGIN = 0x0000, LENGTH = 0x30000 /* 192k */
 }
 
 SECTIONS
